refactor: 💡 change the directory of StringBinaryDitector
