+++ spec/masking/config/target_columns/method/type/string_spec.rb
@@ -22,12 +22,14 @@ RSpec.describe Masking::Config::TargetColumns::Method::Type::String do
 
     # rubocop:disable Style/FormatStringToken
     context 'with sequential number placeholder %{n}' do
+      subject(:subject_object) { described_class.new(value) }
+
       let(:value) { 'number%{n}' }
 
       it 'increment number each by call' do
-        expect(described_class.new(value).call('sql_value')).to eq "'number1'"
-        expect(described_class.new(value).call('sql_value')).to eq "'number2'"
-        expect(described_class.new(value).call('sql_value')).to eq "'number3'"
+        expect(subject_object.call('sql_value')).to eq "'number1'"
+        expect(subject_object.call('sql_value')).to eq "'number2'"
+        expect(subject_object.call('sql_value')).to eq "'number3'"
       end
     end
     # rubocop:enable Style/FormatStringToken
