+++ spec/masking/config/target_columns/method_spec.rb
@@ -28,7 +28,7 @@ RSpec.describe Masking::Config::TargetColumns::Method do
       let(:method) { 'string' }
 
       it do
-        expect(Masking::Config::TargetColumns::Method::Type::StringBinaryDistinctor).to receive(:new).with('string')
+        expect(Masking::Config::TargetColumns::Method::StringBinaryDistinctor).to receive(:new).with('string')
         subject
       end
